Bug fixes
bug fixes, and new enemy

+++ src/typetesto/FastEnemy.java
@@ -8,7 +8,7 @@ public class FastEnemy extends GameObject{
 
 	private Handler handler;
 	
-	public FastEnemy(int x, int y, ID id, Handler handler) {
+	public FastEnemy(float x, float y, ID id, Handler handler) {
 		super(x, y, id);
 		
 		this.handler = handler;
@@ -18,7 +18,7 @@ public class FastEnemy extends GameObject{
 	}
 	
 	public Rectangle getBounds(){
-    	return new Rectangle(x, y, 16, 16);
+    	return new Rectangle((int)x, (int)y, 16, 16);
     }
 
 	public void tick() {
@@ -31,12 +31,12 @@ public class FastEnemy extends GameObject{
 		if (x <= 0 || x >= Game.WIDTH -16) {
 			velX *= -1;
 		}
-		handler.addObject(new Trail(x, y, ID.Trail, Color.cyan, 16, 16, 0.04f, handler));
+		handler.addObject(new Trail((int)x, (int)y, ID.Trail, Color.cyan, 16, 16, 0.04f, handler));
 	}
 
 	public void render(Graphics g) {
 		g.setColor(Color.cyan);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 
 }
